# Remove Encoding of values in char** For Labels