… doc/source/whatsnew/v1.0.0.rst
… pandas/tests/io/json/test_pandas.py