@jbrockmendel 2019-07-26
can you give an example of the datetime inputs that are trouble?