@WillAyd 2019-07-26
Sure. I'm not sure how to best handle this:

```python
>>> df = pd.DataFrame([[1]], columns=[pd.Timestamp("today")])
>>> df.to_json(date_format="iso")
```

While also supporting:

```python
>>> df.columns = df.columns.astype(object)
>>> df.to_json(date_format="iso")
```

In a concise manner. AFAICT in the first example inspecting the object during iteration with `PyDateTime_Check` doesn't recognize it as a datetime whereas when passed in as an object it does