@WillAyd 2019-07-26
Right now the code handles neither and simply provides out the default str, which I guess is epoch for both cases