@jbrockmendel 2019-07-27
> AFAICT in the first example inspecting the object during iteration with PyDateTime_Check doesn't recognize it as a datetime whereas when passed in as an object it does

Off the top of my head, I'd guess that in the non-object case the array you're iterating over is a `ndarray[datetime64[ns]]` and not a DatetimeArray/DatetimeIndex.  As a result, the objects you get when iterating are datetime64 objects, which do not subclass datetime.  The check for that would be `PyObject_TypeCheck(obj, &PyDatetimeArrType_Type)` (at least that's how we do it in `tslibs.util`).