@WillAyd 2019-07-27
Sorry should clarify - so if the columns are all datetimes it appears that `PyTypeNum_ISDATETIME(type_num)` is true (this was in the pre-existing code as well). Otherwise if mixed objects then `PyDateTime_Check` and `PyDate_Check` would work.

Do we have one function that can arbitrary convert objects from either as appropriate to an isoformat that you know of?