@jbrockmendel 2019-07-27
> Do we have one function that can arbitrary convert objects from either as appropriate to an isoformat that you know of?

In python/cython I would probably just use `Timestamp(obj).isoformat()`.  Not sure how big a hassle that would be to do in the c file