@pep8speaks 2019-08-07
Hello @WillAyd! Thanks for updating this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`pandas/tests/io/json/test_pandas.py`](https://github.com/pandas-dev/pandas/blob/7a5fc9cc16db4ba499c6321c365de4d5f7828015/pandas/tests/io/json/test_pandas.py):

> [Line 1021:7](https://github.com/pandas-dev/pandas/blob/7a5fc9cc16db4ba499c6321c365de4d5f7828015/pandas/tests/io/json/test_pandas.py#L1021): [W291](https://duckduckgo.com/?q=pep8%20W291) trailing whitespace
> [Line 1034:9](https://github.com/pandas-dev/pandas/blob/7a5fc9cc16db4ba499c6321c365de4d5f7828015/pandas/tests/io/json/test_pandas.py#L1034): [E265](https://duckduckgo.com/?q=pep8%20E265) block comment should start with '# '
> [Line 1061:9](https://github.com/pandas-dev/pandas/blob/7a5fc9cc16db4ba499c6321c365de4d5f7828015/pandas/tests/io/json/test_pandas.py#L1061): [E265](https://duckduckgo.com/?q=pep8%20E265) block comment should start with '# '