@WillAyd 2019-08-08
I think this is reviewable again. This is a little complicated so to summarize this change it would remove custom label encoding that we do in objToJSON. This theoretically would make the code easier to reason about and opens up the possibility to easily add #12004 indent support through work that is identical or similar to #26457 

A few caveats:
 - This change only affects *labels* which depending on the format could be any combination of the index and column values, if any at all
 - This causes performance regressions dealing with datetimelike objects as *labels only*. Benchmarks and reasoning are showing in Appendix A
 - Previous behavior would recursively encode labels. The plus side of this was that how items appear as keys would use the same logic was what appeared as a value for JSON objects. The downside is that this fights the JSON spec and ujson, as object keys by definition MUST be strings. It is also non-trivial to produce invalid JSON this way (see appendix B)
 - While the ultimate goal is to simplify things, this does introduce slightly different behavior for labels versus the actual data. For example, this change partially solves #15137 by appropriately handling Timedelta data in the labels, but doesn't work for Timedeltas appearing in the data. Ideally how Python objects are handled would be consistent regardless of location, but I've left this inconsistency for now because to truly make it work and resolve all performance implications might take a while

I feel like (save a few minor cleanups) it would be easiest to merge this in the current state and resolve label / data inconsistencies in follow ups, since to do that all at once would be a huge diff and tough to tackle incrementally. Open to suggestions on how to handle though.

## Appendix A
Here are all of the performance regressions. The reason for these is that when detecting a datetime-like array the new label encoding will create Python Timestamp objects for each entry, so that it can appropriately build an isoformat or scale the units from there. 

A vectorized solution would help a lot here and wouldn't be unreasonable to implement. I don't think this will be a long term / architectural problem just needs focus and addressing in follow up

```sh
       before           after         ratio
     [35810733]       [104d2f27]
     <remove-buf-writes~3^2>       <remove-buf-writes>
+        136±30ms        462±100ms     3.39  io.json.ToJSON.time_to_json('index', 'df_int_floats')
+         139±6ms        465±100ms     3.34  io.json.ToJSON.time_to_json('index', 'df_int_float_str')
+        125±30ms         365±80ms     2.93  io.json.ToJSON.time_to_json('index', 'df_date_idx')
+        131±10ms        357±100ms     2.72  io.json.ToJSON.time_to_json('index', 'df_td_int_ts')
+        117±20ms         309±40ms     2.64  io.json.ToJSON.time_to_json('columns', 'df_int_float_str')
+        128±20ms          301±6ms     2.36  io.json.ToJSON.time_to_json('columns', 'df_int_floats')
+        130±20ms         295±40ms     2.27  io.json.ToJSON.time_to_json('columns', 'df_date_idx')
+        150±50ms         301±20ms     2.01  io.json.ToJSON.time_to_json('columns', 'df_td_int_ts')
+         126±5ms         240±10ms     1.91  io.json.ToJSON.time_to_json_wide('records', 'df_date_idx')
+        194±10ms         323±40ms     1.67  io.json.ToJSON.time_to_json_wide('records', 'df_td_int_ts')
+         196±4ms         288±50ms     1.47  io.json.ToJSON.time_to_json_wide('records', 'df_int_float_str')
+         191±2ms         233±20ms     1.22  io.json.ToJSON.time_to_json_wide('columns', 'df_int_float_str')
+         125±1ms         149±10ms     1.19  io.json.ToJSONLines.time_floats_with_int_idex_lines
```

## Appendix B
As mentioned we would previously encode key values for JSON objects. This is pretty easy to break as shown in the example below (note the invalid string that this produces without error):

```python
In [15]: class Foo:
    ...:
    ...:     def toDict(self):
    ...:         return {"a": "b"}
    ...:
    ...: df = pd.DataFrame([[1]], columns=[Foo()])
    ...: print(df.to_json())
{"{"a":"b"}":{"0":1}}
```

This PR instead handles datetimelike objects by converting them to either epoch or iso formats and stringifying from there. All other objects are simply stringified. 

This has the added bonus of making tuples round-trippable albeit with data loss. To illustrate here is current master behavior:

```python
In [18]: pd.read_json(pd.DataFrame([[1]], index=[('a', 'b')]).to_json())
ValueError: No ':' found when decoding object value
```

As opposed to on this branch:
```python
In [5]: pd.read_json(pd.DataFrame([[1]], index=[('a', 'b')]).to_json())
Out[5]:
            0
('a', 'b')  1
```

So again some data loss but at the very least does not produce invalid JSON